@@ -9,31 +9,30 @@ import {
 	Card,
 	FormControlLabel,
 	Switch,
-	IconButton
-} from '@material-ui/core'
+	IconButton,
+} from '@material-ui/core';
 import { PlayArrow, VolumeDown, VolumeUp } from '@material-ui/icons';
 import PauseIcon from '@material-ui/icons/Pause';
-import { SoundHandler } from '../../sound-handler';
+import { SoundHandler } from '../../sound-handler/sound-handler';
 
 export interface SoundCardProps {
-	id: number,
-	cardName: string,
-	soundVolume: number,
-	isCardPlayButtonIsOn: boolean,
-	isCardToneIsOn: boolean,
-	isVisible: boolean,
-	cardIcon: React.ElementType,
-	cardSoundPath: string,
-	isMasterPlayButtonOn: boolean,
-	cardHertzValue: number,
+	id: number;
+	cardName: string;
+	soundVolume: number;
+	isCardPlayButtonIsOn: boolean;
+	isCardToneIsOn: boolean;
+	isVisible: boolean;
+	cardIcon: React.ElementType;
+	cardSoundPath: string;
+	isMasterPlayButtonOn: boolean;
+	cardHertzValue: number;
 
-	onPlayButtonChange: (id: number, isPlaying: boolean) => void,
-	onVolumeChange: (id: number, newValue: number | number[]) => void,
-	onToneSwitchChange: (id: number, isToned: boolean) => void,
+	onPlayButtonChange: (id: number, isPlaying: boolean) => void;
+	onVolumeChange: (id: number, newValue: number | number[]) => void;
+	onToneSwitchChange: (id: number, isToned: boolean) => void;
 }
 
-export interface SoundCardState {
-}
+export interface SoundCardState {}
 
 class SoundCard extends React.Component<SoundCardProps, SoundCardState> {
 	private sound: SoundHandler = new SoundHandler(this.props.cardSoundPath);
@@ -45,51 +44,61 @@ class SoundCard extends React.Component<SoundCardProps, SoundCardState> {
 
 	handleToneSwitch = () => {
 		this.props.onToneSwitchChange(this.props.id, !this.props.isCardToneIsOn);
-	}
+	};
 
 	handlePlayButtonClick = () => {
-		this.props.onPlayButtonChange(this.props.id, !this.props.isCardPlayButtonIsOn);
-	}
+		this.props.onPlayButtonChange(
+			this.props.id,
+			!this.props.isCardPlayButtonIsOn
+		);
+	};
 
 	handleSliderChange = (event: any, newValue: number | number[]) => {
 		this.props.onVolumeChange(this.props.id, newValue);
-		this.sound.setVolume(newValue as number / 50);
-	}
+		this.sound.setVolume((newValue as number) / 50);
+	};
 
 	playSound = () => {
-		if (this.props.isCardPlayButtonIsOn && this.props.isMasterPlayButtonOn && this.props.isVisible) {
+		if (
+			this.props.isCardPlayButtonIsOn &&
+			this.props.isMasterPlayButtonOn &&
+			this.props.isVisible
+		) {
 			this.sound.startSound();
-		}
-		else {
+		} else {
 			this.sound.stopSound();
 		}
-	}
+	};
 
 	playTonedSound = () => {
 		if (this.props.isCardToneIsOn) {
 			this.sound.activateTone(this.props.cardHertzValue);
-		}
-		else {
+		} else {
 			this.sound.deactivateTone();
 		}
-	}
+	};
 
 	render() {
-		const individualPlayButton = this.props.isCardPlayButtonIsOn ? <PauseIcon/> : <PlayArrow/>;
+		const individualPlayButton = this.props.isCardPlayButtonIsOn ? (
+			<PauseIcon />
+		) : (
+			<PlayArrow />
+		);
 		const disablePlayButton = !this.props.isMasterPlayButtonOn;
 		return (
 			<Card>
 				<CardContent>
-					<Grid container spacing={2} alignItems='center' justify='center'>
+					<Grid
+						container
+						spacing={2}
+						alignItems='center'
+						justifyContent='center'
+					>
 						<Grid item>
-							<Typography variant='h4'>
-								{<this.props.cardIcon/>}
-							</Typography>
+							<Typography variant='h4'>{<this.props.cardIcon />}</Typography>
 						</Grid>
 						<Grid item xs>
-							<Typography variant='h6'>
-								{this.props.cardName}
-							</Typography>
+							<Typography variant='h6'>{this.props.cardName}</Typography>
 						</Grid>
 						<Grid item xs>
 							<IconButton
@@ -115,13 +124,16 @@ class SoundCard extends React.Component<SoundCardProps, SoundCardState> {
 							/>
 						</Grid>
 					</Grid>
-					<Typography>
-						Volume
-					</Typography>
+					<Typography>Volume</Typography>
 					<CardActions>
-						<Grid container spacing={2} alignItems='center' justify='center'>
+						<Grid
+							container
+							spacing={2}
+							alignItems='center'
+							justifyContent='center'
+						>
 							<Grid item>
-								<VolumeDown/>
+								<VolumeDown />
 							</Grid>
 							<Grid item xs>
 								<Slider
@@ -134,7 +146,7 @@ class SoundCard extends React.Component<SoundCardProps, SoundCardState> {
 								/>
 							</Grid>
 							<Grid item>
-								<VolumeUp/>
+								<VolumeUp />
 							</Grid>
 						</Grid>
 					</CardActions>
