@@ -7,6 +7,7 @@ export class SoundHandler {
 
 	//Connecting the nodes together
 	constructor(path: string) {
+		console.log(path);
 		this.audio = new Audio(path);
 		this.audio.loop = true;
 		this.context = new AudioContext();
@@ -23,7 +24,11 @@ export class SoundHandler {
 		if (this.context.state === 'suspended') {
 			this.context.resume();
 		}
-		this.audio.play();
+		var promise = this.audio.play();
+		if (promise) {
+			//Older browsers may not return a promise, according to the MDN website
+			promise.catch(function (error) { console.error(error); });
+		}
 	}
 
 	stopSound() {
